Move the kernel application into its own module.
