+++ envisage/plugins/ipython_kernel/internal_ipkernel.py
@@ -10,22 +10,18 @@ import sys
 import ipykernel
 from ipykernel.connect import connect_qtconsole
 from ipykernel.ipkernel import IPythonKernel
-from ipykernel.kernelapp import IPKernelApp
 from ipykernel.zmqshell import ZMQInteractiveShell
 import IPython.utils.io
 import six
 from tornado import ioloop
-import zmq
 
+from envisage.plugins.ipython_kernel.kernelapp import IPKernelApp
 from traits.api import Any, HasStrictTraits, Instance, List
-
-from envisage.plugins.ipython_kernel.heartbeat import Heartbeat
 
 NEEDS_IOLOOP_PATCH = Version(ipykernel.__version__) >= Version('4.7.0')
 
 
-# XXX Refactor to avoid the fragile subclassing of IPKernelApp.
-# XXX Move this class to its own module.
+# XXX Refactor to avoid the fragile subclassing of IPKernelApp. (One day.)
 # XXX Make sure no fds are being leaked *even* if we're keeping all
 #     the relevant objects alive. We shouldn't be depending on reference-count
 #     based cleanup to reclaim fds.
@@ -37,6 +33,8 @@ NEEDS_IOLOOP_PATCH = Version(ipykernel.__version__) >= Version('4.7.0')
 #     attributes *before* save_sys_module_state gets called, and why?
 # XXX Check list of *all* objects for anything suspiciously IPython-like.
 #     What's the delta between the zeroth run and the first?
+# XXX Consider avoiding the "instance" singleton stuff. Maybe save that
+#     for the rewrite.
 
 
 if six.PY2:
@@ -51,180 +49,6 @@ else:
     from atexit import unregister as atexit_unregister
 
 
-class PatchedIPKernelApp(IPKernelApp):
-    """
-    Patched version of the IPKernelApp, mostly to support clean shutdown.
-
-    """
-    # Override to use local Heartbeat class.
-    heartbeat = Instance(Heartbeat, allow_none=True)
-
-    def _unbind_socket(self, s, port):
-        transport = self.transport
-
-        if port <= 0:
-            # This should only happen if _bind_socket hasn't been called.
-            # So we shouldn't ever get here, but perhaps we should defensively
-            # do nothing (maybe print a warning about unbinding a port
-            # that was never bound) if we do get here.
-            raise RuntimeError("Shouldn't ever get here.")
-
-        # XXX Repetition of information from _bind_socket; refactor
-        # to remove the repetition.
-        if transport == "tcp":
-            endpoint = "tcp://%s:%i" % (self.ip, port)
-        elif transport == "ipc":
-            endpoint = "ipc://%s-%i" % (self.ip, port)
-        else:
-            raise ValueError(
-                "Unknown transport type: {}".format(self.transport))
-
-        s.unbind(endpoint)
-
-    def close_sockets(self):
-        """
-        Unbind, close and destroy sockets created by init_sockets.
-        """
-        # context = zmq.Context.instance()
-        self.close_iopub()
-
-        # XXX Already closed by close_kernel, but not explictly unbound.
-        # unbind after close fails (not surprisingly)
-        # Do we still need to unbind?
-        #self._unbind_socket(self.control_socket, self.control_port)
-        #self.control_socket.close()
-        del self.control_socket
-
-        self._unbind_socket(self.stdin_socket, self.stdin_port)
-        self.stdin_socket.close()
-        del self.stdin_socket
-
-        # Also already closed when closing the associated ZMQStream
-        #self._unbind_socket(self.shell_socket, self.shell_port)
-        #self.shell_socket.close()
-        del self.shell_socket
-
-    def close_iopub(self):
-        """
-        Close iopub-related resources.
-        """
-        self.iopub_socket = self.iopub_thread.socket
-
-        self.iopub_thread.stop()
-
-        # Remove the atexit handler that's registered.
-        atexit_unregister(self.iopub_thread.stop)
-
-        # self.iopub_thread._event_puller.stop_on_recv()
-        # XXX Again, not unbinding. Problem?
-        # self.iopub_thread._event_puller.close()
-
-        # self.iopub_thread._pipe_in.stop_on_recv()
-        # XXX Again, not unbinding. Problem?
-        #self.iopub_thread._pipe_in.close()
-
-        # Lots of auxiliary sockets for the IOPubThread. Should we
-        # be closing these before or after ?
-
-        self._unbind_socket(self.iopub_socket, self.iopub_port)
-        self.iopub_socket.close()
-        del self.iopub_socket
-
-    def close_kernel(self):
-        """
-        Undo setup from init_kernel.
-        """
-        # Dig in to find the shell stream and control stream.
-        # kernel is a an IPythonKernel object
-
-        # XXX This is ugly: ideally, we should be closing the streams
-        # in the same class that creates them. That may be awkward
-        # in practice.
-        kernel = self.kernel
-        for stream in kernel.shell_streams:
-            stream.stop_on_recv()
-            # This also closes the corresponding socket.
-            stream.close()
-
-    def close_shell(self):
-        # clean up and close the shell attached to the kernel
-        shell = self.kernel.shell
-
-        # Clean up script magics object, which is contained in a reference
-        # cycle, and has a __del__ method, preventing its removal in Python 2.
-        magics_manager = shell.magics_manager
-        script_magics = magics_manager.registry["ScriptMagics"]
-
-        script_magics.kill_bg_processes()
-        atexit_unregister(script_magics.kill_bg_processes)
-
-        script_magics.magics.clear()
-        script_magics.shell = None
-        script_magics.parent = None
-
-        shell.history_manager.end_session()
-        # The stop also cleans up the file connection.
-        shell.history_manager.save_thread.stop()
-        # Don't run again at process shutdown.
-        atexit_unregister(shell.history_manager.save_thread.stop)
-        # Rely on garbage collection to clean up the file connection.
-        shell.history_manager.db.close()
-        del shell.configurables[:]
-
-        del shell.sys_excepthook
-        del shell._orig_sys_module_state
-        del shell._orig_sys_modules_main_mod
-
-        # atexit_operations does:
-        # - history_manager end session (called above)
-        # - remove temp files and directories
-        # - clear user namespaces
-        # - run user shutdown hooks
-        # Do we want to do all these things? Should we just call
-        # atexit_operations directly?
-        atexit_unregister(shell.atexit_operations)
-
-    def close_heartbeat(self):
-        # This should interrupt the zmq.device call.
-        self.heartbeat.context.term()
-        self.heartbeat.join()
-
-    def init_heartbeat(self):
-        """start the heart beating"""
-        # Overridden from parent to use the local Heartbeat class.
-
-        # heartbeat doesn't share context, because it mustn't be blocked
-        # by the GIL, which is accessed by libzmq when freeing zero-copy messages
-        hb_ctx = zmq.Context()
-        self.heartbeat = Heartbeat(hb_ctx, (self.transport, self.ip, self.hb_port))
-        self.hb_port = self.heartbeat.port
-        self.log.debug("Heartbeat REP Channel on port: %i" % self.hb_port)
-        self.heartbeat.start()
-
-    def XXXXinit_heartbeat(self):
-        # Temporarily override while debugging, to avoid creating the
-        # heartbeat at all. To do: allow the heartbeat to be created,
-        # then shut it down (along with its sockets and context).
-        pass
-
-    def patch_io(self):
-        # Overridden to do nothing. Alternatively, we need to write
-        # and call a corresponding function at teardown that undoes
-        # the faulthandler monkeypatching.
-
-        # Related: https://github.com/ipython/ipykernel/issues/91
-
-        # Suggest that the appropriate fix for users of this class
-        # is to make sure they enable faulthandler before the
-        # IPython kernel does its sys magic.
-        pass
-
-    def configure_tornado_logger(self):
-        # Overridden to do nothing.
-        pass
-
-
-
 def gui_kernel(gui_backend):
     """ Launch and return an IPython kernel GUI with support.
 
@@ -236,7 +60,7 @@ def gui_kernel(gui_backend):
       without GUI support.
     """
 
-    kernel = PatchedIPKernelApp.instance()
+    kernel = IPKernelApp.instance()
 
     argv = ['python']
     if gui_backend is not None:
@@ -251,7 +75,7 @@ class InternalIPKernel(HasStrictTraits):
     """
 
     #: The IPython kernel.
-    ipkernel = Instance(PatchedIPKernelApp)
+    ipkernel = Instance(IPKernelApp)
 
     #: A list of connected Qt consoles.
     consoles = List()
@@ -458,4 +282,4 @@ class InternalIPKernel(HasStrictTraits):
             # Remove singletons.
             ZMQInteractiveShell.clear_instance()
             IPythonKernel.clear_instance()
-            PatchedIPKernelApp.clear_instance()
+            IPKernelApp.clear_instance()
